+++ src/device/woplugmini.ts
@@ -1,12 +1,14 @@
-import { Buffer } from 'buffer';
-
+/* Copyright(C) 2024, donavanbecker (https://github.com/donavanbecker). All rights reserved.
+ *
+ * woplugmini.ts: Switchbot BLE API registration.
+ */
 import { SwitchbotDevice } from '../device.js';
 
 /**
  * @see https://github.com/OpenWonderLabs/SwitchBotAPI-BLE/blob/latest/devicetypes/plugmini.md
  */
 export class WoPlugMini extends SwitchbotDevice {
-  static parseServiceData_US(manufacturerData, onlog) {
+  static parseServiceData_US(manufacturerData: Buffer, onlog: ((message: string) => void) | undefined) {
     if (manufacturerData.length !== 14) {
       if (onlog && typeof onlog === 'function') {
         onlog(
@@ -45,7 +47,7 @@ export class WoPlugMini extends SwitchbotDevice {
     return data;
   }
 
-  static parseServiceData_JP(manufacturerData, onlog) {
+  static parseServiceData_JP(manufacturerData: Buffer, onlog: ((message: string) => void) | undefined) {
     if (manufacturerData.length !== 14) {
       if (onlog && typeof onlog === 'function') {
         onlog(
@@ -94,7 +96,7 @@ export class WoPlugMini extends SwitchbotDevice {
   /**
    * @private
    */
-  _setState(reqByteArray) {
+  _setState(reqByteArray: number[]) {
     const base = [0x57, 0x0f, 0x50, 0x01];
     return this._operateBot([...base, ...reqByteArray]);
   }
@@ -123,7 +125,7 @@ export class WoPlugMini extends SwitchbotDevice {
   /**
    * @private
    */
-  _operateBot(bytes) {
+  _operateBot(bytes: number[]) {
     const req_buf = Buffer.from(bytes);
     return new Promise((resolve, reject) => {
       this._command(req_buf)