+++ src/parameter-checker.ts
@@ -1,5 +1,7 @@
-import { Buffer } from 'buffer';
-
+/* Copyright(C) 2024, donavanbecker (https://github.com/donavanbecker). All rights reserved.
+ *
+ * parameter-checker.ts: Switchbot BLE API registration.
+ */
 type Rule = {
   required?: boolean;
   min?: number;
@@ -8,11 +10,12 @@ type Rule = {
   maxBytes?: number;
   pattern?: RegExp;
   enum?: any[];
+  type?: 'float' | 'integer' | 'boolean' | 'array' | 'object' | 'string';
 };
 
 export class ParameterChecker {
 
-  _error;
+  _error: Error | null;
   static error: any;
   constructor() {
     this._error = null;
@@ -30,7 +33,7 @@ export class ParameterChecker {
     return this._error;
   }
 
-  static isSpecified(value) {
+  static isSpecified(value: unknown) {
     return value === void 0 ? false : true;
   }
 
@@ -67,7 +70,7 @@ export class ParameterChecker {
    *   throw new Error(message);
    * }
    * ---------------------------------------------------------------- */
-  static check(obj, rules, required) {
+  static check(obj: Record<string, unknown>, rules: {[key: string]: Rule}, required: boolean) {
     this.error;
     if (required) {
       if (!this.isSpecified(obj)) {
@@ -97,7 +100,7 @@ export class ParameterChecker {
     for (let i = 0; i < name_list.length; i++) {
       const name = name_list[i];
       const v = obj[name];
-      let rule = rules[name];
+      let rule: Rule = rules[name as keyof typeof rules] as Rule;
 
       if (!rule) {
         rule = {};
@@ -169,7 +172,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isFloat(value, rule: Rule, name = 'value') {
+  static isFloat(value: unknown, rule: Rule, name = 'value'): boolean {
     this.error;
 
     if (!rule.required && !this.isSpecified(value)) {
@@ -251,7 +254,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isInteger(value, rule: Rule, name = 'value') {
+  static isInteger(value: unknown, rule: Rule, name = 'value') {
     this.error = null;
 
     if (!rule.required && !this.isSpecified(value)) {
@@ -259,7 +262,7 @@ export class ParameterChecker {
     }
 
     if (this.isFloat(value, rule)) {
-      if (value % 1 === 0) {
+      if ((value as number) % 1 === 0) {
         return true;
       } else {
         this.error = {
@@ -288,7 +291,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isBoolean(value, rule: Rule, name = 'value') {
+  static isBoolean(value: unknown, rule: Rule, name = 'value') {
     this.error = null;
 
     if (!rule.required && !this.isSpecified(value)) {
@@ -320,7 +323,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isObject(value, rule: Rule, name = 'value') {
+  static isObject(value: unknown, rule: Rule, name = 'value') {
     this.error = null;
     if (!rule.required && !this.isSpecified(value)) {
       return true;
@@ -356,7 +359,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isArray(value, rule: Rule, name = 'value') {
+  static isArray(value: unknown, rule: Rule, name = 'value') {
     this.error = null;
 
     if (!rule.required && !this.isSpecified(value)) {
@@ -427,7 +430,7 @@ export class ParameterChecker {
    * - If the value is invalid, this method will return `false` and
    *   an `Error` object will be set to `this._error`.
    * ---------------------------------------------------------------- */
-  static isString(value, rule: Rule, name = 'value') {
+  static isString(value: unknown, rule: Rule, name = 'value') {
     this.error = null;
 
     if (!rule.required && !this.isSpecified(value)) {